Enhance test cases to increase coverage
Increased function coverage and branch coverage.

Attempted to create a test case for `if (auth?.user)` to be fully evaluated to increase branch coverage. However, it appears that React State change may have some optimization, causing the `setName` `setPhone` `setAddress` to be optimized and not covered by the tests. Hence the coverage report may indicate that these lines are not covered.